Fix type

+++ src/libraries/System.IO.Compression.Zstandard/src/System/IO/Compression/ZstandardEncoder.cs
@@ -262,7 +262,7 @@ namespace System.IO.Compression
                 throw new ArgumentOutOfRangeException(nameof(inputSize));
             }
 
-            if (result > int.MaxValue)
+            if (result > long.MaxValue)
             {
                 throw new ArgumentOutOfRangeException(nameof(inputSize));
             }
